refactor code duplicatoin
